+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.2.3 (TBA)
+
+* Remove `:castore` version requirement
+* `Assent.Strategy.Httpc.request/5` raises error when SSL certificate can't be validated
+
 ## v0.2.2 (2023-02-27)
 
 * Fixed bug to handle 201 success response